Move gate_product to circuit/simulator/utils.py
