+++ src/qutip_qip/circuit/simulator/__init__.py
@@ -1,6 +1,6 @@
 from .result import CircuitResult
-from .circuitsimulator import (
-    CircuitSimulator,
+from .matrix_mul_simulator import CircuitSimulator
+from .utils import (
     gate_sequence_product,
     gate_sequence_product_with_expansion,
 )
